fix: reusable types for HyGraph
